grpcserver use different datastore based on config
